Midway Through adapting Salt_Sim for mobile.
Doing a quick save and revert so the live website looks presentable
