+++ src/Menu.css
@@ -220,7 +220,7 @@
     justify-self: right;
     margin: 0px 0px 0px 0px ;
     position: relative;
-    border-radius: 30px 30px 0px 0px;
+    border-radius: 30px 30px 30px 30px;
     padding-left: 2%;
     padding-right: 2%;
     top: -40vh;
@@ -237,7 +237,7 @@
 
 
     transition: transform 0.5s ease;
-    transform-origin: bottom;
+    transform-origin: top left;
   }
 
   .network-menu-wrapper.collapsed {
@@ -246,7 +246,7 @@
   }
 
   .network-menu-wrapper.open {
-    transform: scale(0.55);
+    transform: scale(1);
   }
