+++ src/Salt_Sim_Files/NetworkMenu.js
@@ -43,9 +43,11 @@ const Network_Menu = (props) => {
 
     <div 
         className={`network-menu-wrapper ${networkMenuOpen ? 'open' : 'collapsed'}`}
-        style={{ pointerEvents: 'auto' }}>
+        style={{ pointerEvents: 'none', position: 'relative', left: '0%', bottom: '0%', width: '100%', height: '100%', zIndex: '100'}}>
 
-        <networkmenu>
+        <networkmenu style={{ pointerEvents: 'auto', position: 'absolute', left: '50%', top: '5%',
+                   overflowY: 'auto', padding: '10px', width: '50%', minWidth: '30px', maxHeight: '160%',
+                   transform: 'scale(0.55)', transformOrigin: 'top left'}}> 
 
         <h1 >   Translation is Liquid  </h1>
 
@@ -87,8 +89,8 @@ const Network_Menu = (props) => {
             
             src={play_button_lever_img}/>
 
-            <div style = {{position:'relative', left: '55%', top: '-55%'}}>On</div>
-            <div style = {{position:'relative', left: '110%', top: '0%'}}>Off</div>
+            <div style = {{position:'relative', left: '92%', top: '-55%'}}>On</div>
+            <div style = {{position:'relative', left: '147%', top: '0%'}}>Off</div>
