+++ src/Salt_Sim_Files/SaveLoadButton.css
@@ -31,8 +31,8 @@
     width: 100%;
     height: 15%;
     position: absolute;
-    top: 75%;
-    left: 0%;
+    top: 70%;
+    left: 0.5%;
     
     background-color: rgb(21, 45, 82);
     border: 1px solid #ccc;
